+++ reference/src/srd_5e/mod.rs
@@ -1,4 +1,4 @@
-pub use equipment::{Item, ItemCategory};
+pub use equipment::{Item, ItemCategory, MagicItem};
 pub use spell::Spell;
 pub use std::fmt;
 
@@ -26,6 +26,13 @@ pub fn item_categories() -> Result<Vec<ItemCategory>, String> {
     .map_err(|e| format!("{}", e))
 }
 
+pub fn magic_items() -> Result<Vec<MagicItem>, String> {
+    serde_json::from_str(include_str!(
+        "../../../data/srd_5e/src/5e-SRD-Magic-Items.json",
+    ))
+    .map_err(|e| format!("{}", e))
+}
+
 #[derive(Debug, Deserialize)]
 pub struct Reference {
     index: String,
@@ -42,11 +49,11 @@ impl Reference {
 fn write_text_block(f: &mut fmt::Formatter, lines: &[String]) -> fmt::Result {
     let mut prev_line: Option<&str> = None;
 
+    let is_list = |l: &str| l.starts_with("- ") || l.starts_with("* ") || l.starts_with('|');
+
     for line in lines.iter() {
         if prev_line.is_some() {
-            if !prev_line.map_or(false, |l| l.starts_with(&['-', '*'][..]))
-                || !line.starts_with(&['-', '*'][..])
-            {
+            if !prev_line.map_or(false, is_list) || !is_list(line) {
                 writeln!(f)?;
             }
             writeln!(f)?;